+++ frontend/pages/UserSettingsPage/UserSettingsPage.tsx
@@ -1,36 +1,33 @@
-import React, { useState, useContext, useEffect, ReactElement } from "react";
+import React, { useState, useContext } from "react";
 import { InjectedRouter } from "react-router";
-import { formatDistanceToNow } from "date-fns";
 import classnames from "classnames";
 
 import { AppContext } from "context/app";
 import { NotificationContext } from "context/notification";
-import { IVersionData } from "interfaces/version";
 import { IUser } from "interfaces/user";
 import usersAPI from "services/entities/users";
-import versionAPI from "services/entities/version";
 import { authToken } from "utilities/local";
-import { stringToClipboard } from "utilities/copy_text";
 import deepDifference from "utilities/deep_difference";
 import formatErrorResponse from "utilities/format_error_response";
-import { generateRole, generateTeam, greyCell } from "utilities/helpers";
 
-import Avatar from "components/Avatar";
 import Button from "components/buttons/Button";
 // @ts-ignore
 import ChangeEmailForm from "components/forms/ChangeEmailForm";
 // @ts-ignore
 import ChangePasswordForm from "components/forms/ChangePasswordForm";
 // @ts-ignore
-import FleetIcon from "components/icons/FleetIcon";
-// @ts-ignore
 import Modal from "components/Modal";
+
 // @ts-ignore
 import UserSettingsForm from "components/forms/UserSettingsForm";
 import InfoBanner from "components/InfoBanner";
 import SecretField from "components/SecretField";
 import SandboxGate from "components/SandboxGate";
+import MainContent from "components/MainContent";
+import SidePanelContent from "components/SidePanelContent";
+
 import ExternalURLIcon from "../../../assets/images/icon-external-url-12x12@2x.png";
+import UserSidePanel from "./UserSidePanel";
 
 const baseClass = "user-settings";
 
@@ -41,35 +38,16 @@ interface IUserSettingsPageProps {
 const UserSettingsPage = ({
   router,
 }: IUserSettingsPageProps): JSX.Element | null => {
-  const { config, currentUser, isPremiumTier, isSandboxMode } = useContext(
-    AppContext
-  );
+  const { config, currentUser, isSandboxMode } = useContext(AppContext);
   const { renderFlash } = useContext(NotificationContext);
 
   const [pendingEmail, setPendingEmail] = useState<string>("");
   const [showEmailModal, setShowEmailModal] = useState<boolean>(false);
   const [showPasswordModal, setShowPasswordModal] = useState<boolean>(false);
   const [updatedUser, setUpdatedUser] = useState<Partial<IUser>>({});
-  const [copyMessage, setCopyMessage] = useState<string>("");
   const [showApiTokenModal, setShowApiTokenModal] = useState<boolean>(false);
-  const [revealSecret, setRevealSecret] = useState<boolean>(false);
-  const [versionData, setVersionData] = useState<IVersionData>();
   const [errors, setErrors] = useState<{ [key: string]: string }>({});
   const [userErrors, setUserErrors] = useState<{ [key: string]: string }>({});
-
-  useEffect(() => {
-    const getVersionData = async () => {
-      try {
-        const data = await versionAPI.load();
-        setVersionData(data);
-      } catch (response) {
-        console.error(response);
-        return false;
-      }
-    };
-
-    getVersionData();
-  }, []);
 
   const onCancel = (evt: React.MouseEvent<HTMLButtonElement>) => {
     evt.preventDefault();
@@ -100,21 +78,6 @@ const UserSettingsPage = ({
   const onToggleApiTokenModal = () => {
     setShowApiTokenModal(!showApiTokenModal);
     return false;
-  };
-
-  // placeholder is needed even though it's not used
-  const onCopySecret = (placeholder: string) => {
-    return (evt: ClipboardEvent) => {
-      evt.preventDefault();
-
-      stringToClipboard(authToken())
-        .then(() => setCopyMessage("Copied!"))
-        .catch(() => setCopyMessage("Copy failed"));
-
-      // Clear message after 1 second
-      setTimeout(() => setCopyMessage(""), 1000);
-      return false;
-    };
   };
 
   const handleSubmit = async (formData: any) => {
@@ -209,23 +172,6 @@ const UserSettingsPage = ({
     );
   };
 
-  const renderLabel = () => {
-    return (
-      <span className={`${baseClass}__name`}>
-        <span className="buttons">
-          {copyMessage && <span>{`${copyMessage} `}</span>}
-          <Button
-            variant="unstyled"
-            className={`${baseClass}__secret-copy-icon`}
-            onClick={onCopySecret(`.${baseClass}__secret-input`)}
-          >
-            <FleetIcon name="clipboard" />
-          </Button>
-        </span>
-      </span>
-    );
-  };
-
   const renderApiTokenModal = () => {
     if (!showApiTokenModal) {
       return false;
@@ -289,109 +235,47 @@ const UserSettingsPage = ({
     return null;
   }
 
-  const {
-    global_role: globalRole,
-    updated_at: updatedAt,
-    sso_enabled: ssoEnabled,
-    teams,
-  } = currentUser;
-
-  const roleText = generateRole(teams, globalRole);
-  const teamsText = generateTeam(teams, globalRole);
-
-  const lastUpdatedAt =
-    updatedAt &&
-    formatDistanceToNow(new Date(updatedAt), {
-      addSuffix: true,
-    });
-
   const wrapperStyles = classnames(baseClass, {
     [`${baseClass}__sandboxMode`]: isSandboxMode,
   });
 
   return (
-    <div className={wrapperStyles}>
-      <SandboxGate
-        message="Account management is only available in self-managed Fleet"
-        utmSource="fleet-ui-my-account-page"
-      >
-        <div className={`${baseClass}__manage body-wrap`}>
-          <h1>My account</h1>
-          <UserSettingsForm
-            formData={currentUser}
-            handleSubmit={handleSubmit}
-            onCancel={onCancel}
-            pendingEmail={pendingEmail}
-            serverErrors={errors}
-            smtpConfigured={config?.smtp_settings.configured}
-          />
-        </div>
-        <div className={`${baseClass}__additional body-wrap`}>
-          <div className={`${baseClass}__change-avatar`}>
-            <Avatar user={currentUser} className={`${baseClass}__avatar`} />
-            <a href="http://en.gravatar.com/emails/">
-              Change photo at Gravatar
-            </a>
-          </div>
-          {isPremiumTier && (
-            <div className={`${baseClass}__more-info-detail`}>
-              <p className={`${baseClass}__header`}>Teams</p>
-              <p
-                className={`${baseClass}__description ${baseClass}__teams ${greyCell(
-                  teamsText
-                )}`}
-              >
-                {teamsText}
-              </p>
+    <>
+      <MainContent>
+        <div className={wrapperStyles}>
+          <SandboxGate
+            message="Account management is only available in self-managed Fleet"
+            utmSource="fleet-ui-my-account-page"
+          >
+            <div className={`${baseClass}__manage`}>
+              <h1>My account</h1>
+              <UserSettingsForm
+                formData={currentUser}
+                handleSubmit={handleSubmit}
+                onCancel={onCancel}
+                pendingEmail={pendingEmail}
+                serverErrors={errors}
+                smtpConfigured={config?.smtp_settings.configured}
+              />
             </div>
-          )}
-          <div className={`${baseClass}__more-info-detail`}>
-            <p className={`${baseClass}__header`}>Role</p>
-            <p
-              className={`${baseClass}__description ${baseClass}__role ${greyCell(
-                roleText
-              )}`}
-            >
-              {roleText}
-            </p>
-          </div>
-          <div className={`${baseClass}__more-info-detail`}>
-            <p className={`${baseClass}__header`}>Password</p>
-          </div>
-          <Button
-            onClick={onShowPasswordModal}
-            disabled={ssoEnabled}
-            className={`${baseClass}__button`}
-          >
-            Change password
-          </Button>
-          <p className={`${baseClass}__last-updated`}>
-            Last changed: {lastUpdatedAt}
-          </p>
-          <Button
-            onClick={onShowApiTokenModal}
-            className={`${baseClass}__button`}
-          >
-            Get API token
-          </Button>
-          <span
-            className={`${baseClass}__version`}
-          >{`Fleet ${versionData?.version} • Go ${versionData?.go_version}`}</span>
-          <span className={`${baseClass}__privacy-policy`}>
-            <a
-              href="https://fleetdm.com/legal/privacy"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              Privacy policy
-            </a>
-          </span>
+            {renderEmailModal()}
+            {renderPasswordModal()}
+            {renderApiTokenModal()}
+          </SandboxGate>
         </div>
-        {renderEmailModal()}
-        {renderPasswordModal()}
-        {renderApiTokenModal()}
-      </SandboxGate>
-    </div>
+      </MainContent>
+
+      {/* TODO: change after SandboxGate extension */}
+      {!isSandboxMode && (
+        <SidePanelContent>
+          <UserSidePanel
+            currentUser={currentUser}
+            onChangePassword={onShowPasswordModal}
+            onGetApiToken={onShowApiTokenModal}
+          />
+        </SidePanelContent>
+      )}
+    </>
   );
 };
