+++ frontend/pages/UserSettingsPage/UserSidePanel/_styles.scss
@@ -1,0 +1,94 @@
+.user-side-panel {
+
+  &__change-avatar {
+    position: relative;
+    padding: 0 0 20px;
+
+    a {
+      display: block;
+      color: $core-vibrant-blue;
+      text-align: center;
+      position: relative;
+      z-index: 1;
+      text-decoration: none;
+      font-size: $x-small;
+      margin-top: $pad-xsmall;
+    }
+  }
+
+  &__header {
+    font-size: $x-small;
+    font-weight: $bold;
+    margin: $pad-large 0 $pad-xsmall;
+  }
+
+  &__description {
+    font-size: $x-small;
+    font-style: italic;
+    margin: $pad-xsmall 0;
+  }
+
+  &__avatar {
+    margin: 0 auto;
+  }
+
+  &__more-info-detail {
+    color: $core-fleet-black;
+
+    .fleeticon {
+      color: $ui-gray;
+      font-size: 18px;
+      margin-right: 11px;
+    }
+  }
+
+  &__button {
+    width: 100%;
+  }
+
+  &__last-updated {
+    font-size: $x-small;
+    font-style: italic;
+    margin: 4px 0 $pad-large;
+  }
+
+  &__confirm-update {
+    margin-bottom: $pad-medium;
+  }
+
+  &__secret-label {
+    font-size: $x-small;
+    font-weight: $bold;
+    margin: 0 0 $pad-xsmall;
+
+    a {
+      color: $core-vibrant-blue;
+    }
+  }
+
+  &__button-wrap {
+    display: flex;
+    justify-content: flex-end;
+  }
+
+  &__version {
+    font-size: $xx-small;
+    text-align: center;
+    display: block;
+    margin-top: $pad-large;
+  }
+
+  &__privacy-policy {
+    text-align: center;
+    display: block;
+    margin-top: $pad-xlarge;
+
+    a {
+      font-size: $x-small;
+      color: $core-vibrant-blue;
+      font-weight: $bold;
+      text-decoration: none;
+    }
+  }
+
+}
